Refactor: Layout 컴포넌트 적용 - GlobalNav의 파일 위치를 MainLayout으로 이동 - Story, RecommendUser의 위치를 layout 폴더에서 page폴더 내 Home으로 이동 - 반응형 디자인 적용
